refactor: isolate test-only helpers
