+++ web-client/test/uiStore.integration.test.tsx
@@ -4,7 +4,7 @@ import type { Root } from "react-dom/client";
 
 import { runtimeEventHub } from "@client/src/runtime/event-hub";
 import services from "@client/src/runtime/service-registry";
-import { resetUiStoreForTesting, uiStore } from "../src/ui/store";
+import { resetUiStoreForTesting, uiStore } from "./utils/uiStoreTestUtils";
 import CharState from "../src/CharState";
 import GuildsSettings from "../src/options/GuildsSettings";
 import guilds from "../src/options/guilds";
